Add ability to handle a board whose coordinates don't start at 0,0
After I made the assertion based test pass with what I thought would work,
I ran the approvals test, which revealed a contradiction in how the board
was laid out. I'm not sure if the assertion test helped me here or not.
It helped get the initial logic of integrating the BoardBounds set up, so I
didn't have to think about setting that up when I found the contradiction.
That made the things I needed to build smaller.