Fix BoardBounds not being properly calculated for grids larger than 4x4
This was caught with a fully covering approval test on BoardOutputter. Looking
at the test on BoardBounds, this was for two reasons. First, the approval is
not actually applying full coverage. It is essentially a set of assertion
tests. Second, there is no way to verify that the output of BoardBounds is
actually correct without it being included in the context of the
BoardOutputter. In order to address this, I'm moving BoardBounds to be an
inner class of BoardOutputter.