Output RelationshipClassifier test as board
I didn't like the way my map was outputting to my
approved file, because there wasn't a way to see
if it was correct at a glance. I came up with the
idea that it should output as a board string, not
a map output. That required me to make changes
to the existing BoardOutputter to enable it to
create boards in negative board space and create
boards that were larger than 5x5. I needed to make
the changes anyway in order for this to be a
complete GOL solution.

This commit is me modifying the current "SELF"
classification test to use the new BoardOutputter.