Two Neighbour approval working for LiveCellsGraph
It's not pretty. CombinationApprovals just
wouldn't work for this, because the function I'd
need to call would require a List of Cells, and
CombinationApprovals only takes arrays. The bigger
 issue was that I needed to create a set of all
combinations of two neighbour cells.
CombinationsApprovals would approve each result as
it was generated, and because it is a series of
nested for loops... I think I would be evaluating
permutations instead of combinations
(this requires me to double-check the code... I
can't remember). Furthermore,
CombinationsApprovals makes it very difficult to
use the same function to output a result for
combinations of arrays of various sizes... I'm not
 sure how much call there is for this in the "real
 world" though, and TBH, it makes me feel like a
more javascripty-idea that we should be able to do
 that. Perhaps that is the wrong thought.